chore/FE: Reorganize components folder structure
