+++ Front-end/src/app/components/cart/cart-products.tsx
