+++ Front-end/src/app/components/cart/dialogs/cart-erase-product-dialog.tsx
