+++ Front-end/src/app/components/product/dialogs/add-new-category-dialog.tsx
