+++ Front-end/src/app/components/product/dialogs/delete-images-dialog.tsx
