+++ Front-end/src/app/components/product/dialogs/product-post-success-dialog.tsx
