+++ Front-end/src/app/components/product/mobile-product.images.tsx
