+++ Front-end/src/app/components/product/product-layout-toggle.tsx
