+++ Front-end/src/app/products/[id]/page.tsx
@@ -1,6 +1,6 @@
 import ProductImages from "@/app/components/ProductImages";
 import { getProduct } from "@/app/lib/data";
-import ProductForm from "@/app/components/ui/product/product-form";
+import ProductForm from "@/app/components/product/product-form";
 
 export default async function ProductDetails({
   params,
